feat : 댓글 조회

+++ server/Recruit-Api/src/main/java/com/econovation/recruit/application/service/InterviewersService.java
@@ -28,7 +28,7 @@ public class InterviewersService implements InterviewerUseCase {
     }
 
     @Override
-    public Interviewer getById(Integer idpId) {
+    public Interviewer getById(Long idpId) {
         return interviewerLoadPort.loadInterviewById(idpId);
     }
 }
