+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/label/Label.java
@@ -21,5 +21,5 @@ public class Label {
     private Integer applicantId;
 
     @Column(name = "idpId")
-    private Integer idpId;
+    private Long idpId;
 }
